+++ src/Plainion.Windows/Controls/Text/AutoCorrection/BulletList.cs
@@ -8,7 +8,7 @@ namespace Plainion.Windows.Controls.Text.AutoCorrection
         {
             var result = new AutoCorrectionResult(false);
 
-            foreach (var line in DocumentOperations.GetLines(input.Range))
+            foreach (var line in input.Context.GetLines())
             {
                 if (line.Text.StartsWith("- "))
                 {