+++ src/Plainion.Windows/Controls/Text/AutoCorrection/UnicodeSymbolCorrection.cs
@@ -31,7 +31,7 @@ namespace Plainion.Windows.Controls.Text.AutoCorrection
         {
             bool success = false;
 
-            foreach (var wordRange in DocumentOperations.GetWords(input.Range))
+            foreach (var wordRange in input.Context.GetWords())
             {
                 var symbol = Symbols.FirstOrDefault(x => x.Ascii == wordRange.TextOnly());
                 if (symbol != null)